use same action for checking out head
