+++ src/preview.ts
@@ -41,9 +41,12 @@ async function main() {
       throw new Error("github_token is required to make a comment");
     }
 
-    // If we came from the checkout-base action, we might need to save the
+    // If we came from the checkout-pr-ref action, we might need to save the
     // generated config files.
     const { savedSha } = await saveConfig({ oasPath, configPath });
+    if (savedSha !== null && savedSha !== headSha) {
+      throw new Error(`Expected HEAD to be ${headSha}, but was ${savedSha}`);
+    }
 
     const stainless = new Stainless({
       project: projectName,
@@ -54,12 +57,6 @@ async function main() {
     startGroup("Getting parent revision");
 
     const { mergeBaseSha } = await getMergeBase({ baseSha, headSha });
-    if (savedSha !== null && savedSha !== mergeBaseSha) {
-      throw new Error(
-        `Expected HEAD to be ${mergeBaseSha}, but was ${savedSha}`,
-      );
-    }
-
     const { nonMainBaseRef } = await getNonMainBaseRef({
       baseRef,
       defaultBranch,
